Fix oversight

+++ src/data_access/FileUserDataAccessObject.java
@@ -62,7 +62,7 @@ public class FileUserDataAccessObject implements GetRecipeDataAccessInterface, M
                                 Integer.parseInt(details[1]),
                                 Integer.parseInt(details[2]),
                                 Integer.parseInt(details[3]),
-                                Integer.parseInt(details[4]));
+                                Float.parseFloat(details[4]));
                         user.addItem(newItem);
                     }
